+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All changes to this framework is documented here in chronological order (latest
 first).
 
 - 1.0.20
-  - Fix `.frameworkVersion` returning nil when using CocoaPods.
+  - `.frameworkVersion` now returns the version string (previously a optional
+    string).
   - Add an initial implementation of `SwiftyZeroMQ.Poller` with tests (Pull
     request #1) by [Jonathan Cockayne](https://github.com/jcockayne). This is
     based on the [pyzmq](https://github.com/zeromq/pyzmq) implementation.